+++ zcash_primitives/src/sapling/address.rs
@@ -1,0 +1,118 @@
+use group::{Group, GroupEncoding};
+
+use super::{
+    keys::Diversifier,
+    note::{Note, Rseed},
+};
+
+/// A Sapling payment address.
+///
+/// # Invariants
+///
+/// `pk_d` is guaranteed to be prime-order (i.e. in the prime-order subgroup of Jubjub,
+/// and not the identity).
+#[derive(Clone, Debug)]
+pub struct PaymentAddress {
+    pk_d: jubjub::SubgroupPoint,
+    diversifier: Diversifier,
+}
+
+impl PartialEq for PaymentAddress {
+    fn eq(&self, other: &Self) -> bool {
+        self.pk_d == other.pk_d && self.diversifier == other.diversifier
+    }
+}
+
+impl Eq for PaymentAddress {}
+
+impl PaymentAddress {
+    /// Constructs a PaymentAddress from a diversifier and a Jubjub point.
+    ///
+    /// Returns None if `pk_d` is the identity.
+    pub fn from_parts(diversifier: Diversifier, pk_d: jubjub::SubgroupPoint) -> Option<Self> {
+        if pk_d.is_identity().into() {
+            None
+        } else {
+            Some(PaymentAddress { pk_d, diversifier })
+        }
+    }
+
+    /// Constructs a PaymentAddress from a diversifier and a Jubjub point.
+    ///
+    /// Only for test code, as this explicitly bypasses the invariant.
+    #[cfg(test)]
+    pub(crate) fn from_parts_unchecked(
+        diversifier: Diversifier,
+        pk_d: jubjub::SubgroupPoint,
+    ) -> Self {
+        PaymentAddress { pk_d, diversifier }
+    }
+
+    /// Parses a PaymentAddress from bytes.
+    pub fn from_bytes(bytes: &[u8; 43]) -> Option<Self> {
+        let diversifier = {
+            let mut tmp = [0; 11];
+            tmp.copy_from_slice(&bytes[0..11]);
+            Diversifier(tmp)
+        };
+        // Check that the diversifier is valid
+        diversifier.g_d()?;
+
+        let pk_d = jubjub::SubgroupPoint::from_bytes(bytes[11..43].try_into().unwrap());
+        if pk_d.is_some().into() {
+            PaymentAddress::from_parts(diversifier, pk_d.unwrap())
+        } else {
+            None
+        }
+    }
+
+    /// Returns the byte encoding of this `PaymentAddress`.
+    pub fn to_bytes(&self) -> [u8; 43] {
+        let mut bytes = [0; 43];
+        bytes[0..11].copy_from_slice(&self.diversifier.0);
+        bytes[11..].copy_from_slice(&self.pk_d.to_bytes());
+        bytes
+    }
+
+    /// Returns the [`Diversifier`] for this `PaymentAddress`.
+    pub fn diversifier(&self) -> &Diversifier {
+        &self.diversifier
+    }
+
+    /// Returns `pk_d` for this `PaymentAddress`.
+    pub fn pk_d(&self) -> &jubjub::SubgroupPoint {
+        &self.pk_d
+    }
+
+    pub fn g_d(&self) -> Option<jubjub::SubgroupPoint> {
+        self.diversifier.g_d()
+    }
+
+    pub fn create_note(&self, value: u64, rseed: Rseed) -> Option<Note> {
+        self.g_d().map(|g_d| Note {
+            value,
+            rseed,
+            g_d,
+            pk_d: self.pk_d,
+        })
+    }
+}
+
+#[cfg(any(test, feature = "test-dependencies"))]
+pub(super) mod testing {
+    use proptest::prelude::*;
+
+    use super::{
+        super::keys::{testing::arb_incoming_viewing_key, Diversifier, SaplingIvk},
+        PaymentAddress,
+    };
+
+    pub fn arb_payment_address() -> impl Strategy<Value = PaymentAddress> {
+        arb_incoming_viewing_key().prop_flat_map(|ivk: SaplingIvk| {
+            any::<[u8; 11]>().prop_filter_map(
+                "Sampled diversifier must generate a valid Sapling payment address.",
+                move |d| ivk.to_payment_address(Diversifier(d)),
+            )
+        })
+    }
+}
