+++ zcash_primitives/src/sapling/keys.rs
@@ -6,15 +6,16 @@
 
 use std::io::{self, Read, Write};
 
+use super::{address::PaymentAddress, group_hash::group_hash};
 use crate::{
-    constants::{PROOF_GENERATION_KEY_GENERATOR, SPENDING_KEY_GENERATOR},
+    constants::{self, PROOF_GENERATION_KEY_GENERATOR, SPENDING_KEY_GENERATOR},
     keys::prf_expand,
 };
+
+use blake2s_simd::Params as Blake2sParams;
 use ff::PrimeField;
 use group::{Group, GroupEncoding};
 use subtle::CtOption;
-
-use super::{NullifierDerivingKey, ProofGenerationKey, ViewingKey};
 
 /// Errors that can occur in the decoding of Sapling spending keys.
 pub enum DecodingError {
@@ -105,6 +106,60 @@ impl ExpandedSpendingKey {
     }
 }
 
+#[derive(Clone)]
+pub struct ProofGenerationKey {
+    pub ak: jubjub::SubgroupPoint,
+    pub nsk: jubjub::Fr,
+}
+
+impl ProofGenerationKey {
+    pub fn to_viewing_key(&self) -> ViewingKey {
+        ViewingKey {
+            ak: self.ak,
+            nk: NullifierDerivingKey(constants::PROOF_GENERATION_KEY_GENERATOR * self.nsk),
+        }
+    }
+}
+
+/// A key used to derive the nullifier for a Sapling note.
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub struct NullifierDerivingKey(pub jubjub::SubgroupPoint);
+
+#[derive(Debug, Clone)]
+pub struct ViewingKey {
+    pub ak: jubjub::SubgroupPoint,
+    pub nk: NullifierDerivingKey,
+}
+
+impl ViewingKey {
+    pub fn rk(&self, ar: jubjub::Fr) -> jubjub::SubgroupPoint {
+        self.ak + constants::SPENDING_KEY_GENERATOR * ar
+    }
+
+    pub fn ivk(&self) -> SaplingIvk {
+        let mut h = [0; 32];
+        h.copy_from_slice(
+            Blake2sParams::new()
+                .hash_length(32)
+                .personal(constants::CRH_IVK_PERSONALIZATION)
+                .to_state()
+                .update(&self.ak.to_bytes())
+                .update(&self.nk.0.to_bytes())
+                .finalize()
+                .as_bytes(),
+        );
+
+        // Drop the most significant five bits, so it can be interpreted as a scalar.
+        h[31] &= 0b0000_0111;
+
+        SaplingIvk(jubjub::Fr::from_repr(h).unwrap())
+    }
+
+    pub fn to_payment_address(&self, diversifier: Diversifier) -> Option<PaymentAddress> {
+        self.ivk().to_payment_address(diversifier)
+    }
+}
+
 /// A Sapling key that provides the capability to view incoming and outgoing transactions.
 #[derive(Debug)]
 pub struct FullViewingKey {
@@ -186,13 +241,39 @@ impl FullViewingKey {
     }
 }
 
+#[derive(Debug, Clone)]
+pub struct SaplingIvk(pub jubjub::Fr);
+
+impl SaplingIvk {
+    pub fn to_payment_address(&self, diversifier: Diversifier) -> Option<PaymentAddress> {
+        diversifier.g_d().and_then(|g_d| {
+            let pk_d = g_d * self.0;
+
+            PaymentAddress::from_parts(diversifier, pk_d)
+        })
+    }
+
+    pub fn to_repr(&self) -> [u8; 32] {
+        self.0.to_repr()
+    }
+}
+
+#[derive(Copy, Clone, Debug, PartialEq, Eq)]
+pub struct Diversifier(pub [u8; 11]);
+
+impl Diversifier {
+    pub fn g_d(&self) -> Option<jubjub::SubgroupPoint> {
+        group_hash(&self.0, constants::KEY_DIVERSIFICATION_PERSONALIZATION)
+    }
+}
+
 #[cfg(any(test, feature = "test-dependencies"))]
 pub mod testing {
     use proptest::collection::vec;
     use proptest::prelude::*;
     use std::fmt::{self, Debug, Formatter};
 
-    use super::{ExpandedSpendingKey, FullViewingKey};
+    use super::{ExpandedSpendingKey, FullViewingKey, SaplingIvk};
 
     impl Debug for ExpandedSpendingKey {
         fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
@@ -209,6 +290,12 @@ pub mod testing {
     prop_compose! {
         pub fn arb_full_viewing_key()(sk in arb_expanded_spending_key()) -> FullViewingKey {
             FullViewingKey::from_expanded_spending_key(&sk)
+        }
+    }
+
+    prop_compose! {
+        pub fn arb_incoming_viewing_key()(fvk in arb_full_viewing_key()) -> SaplingIvk {
+            fvk.vk.ivk()
         }
     }
 }
